refactor: Co-Locate Styles for Tasks Components
