--- components/styles/TasksStyles.tsx
@@ -1,11 +1,0 @@
-import styled, { css } from "styled-components";
-
-const TasksStyles = styled.div`
-  display: grid;
-  grid-template-columns: 200px 200px;
-  grid-gap: 32px;
-  @media only screen and (max-width: 900px) {
-  }
-`;
-
-export default TasksStyles;
